iot_monitoring_site/requirements.txt
Django==3.0.7
django-cors-headers==3.2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
drf-nested-routers==0.91
